Initial sphinxdocs setup
```
make html
```
The ``_dontscan`` directory contains rst files not hooked in yet because
there **are** plenty of warnings and errors in the sourcecode at the
moment.
:see_no_evil: REM in make.bat to dump this all to a log also for easier
fixins of the actual source code.

This is a basic start to generate all the apidocs with sphinx.
work on it with the refactoring to make a errorfree and **READABLE** set
of documentation that is mobile friendly and is common sight to
pythonistas everywhere.

Zen: Readability Counts!