Add LOOT API install script:
Downloads and installs the API python wrapper from Bintray at the
moment: following LOOT v0.10.0's release this will be updated to
download from GitHub instead.

- Update LOOT API installation script

It now downloads and installs the MSVC 2015 redistributable if
necessary, and saves downloaded files to the OS temporary directory.

- Wrap install script registry access in try/catch

If access fails (eg. because the key doesn't exist) the functions throw,
so assume in that case that the runtime needs to be installed.