Import loot_api without modifying sys.path:
This is necessary for standalone packaging and also makes more sense as
loot_api is an external module so should be loaded like other external
modules.

With this change, loot_api.dll/loot_api.pyd need to be in the Mopy
folder to be found (or presumably wherever PyPI packages are installed
to, but that is untested).

- Update LOOT API installer for new destination

- Fix standalone exe builds

For the executable to load the LOOT API without crashing, the pyd needs
to be inside the executable and the DLL needs to be in the same
directory as the executable.